+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 - Create trait `CustomQuery` for the generic argument in
   `QueryRequest<C: CustomQuery>`. This allows us to provide
   `impl<C: CustomQuery> From<C> for QueryRequest<C>` for any custom query.
+- Implement `From<Binary> for Vec<u8>`.
 
 **cosmwasm-storage**